refactor: Modify files structure
