+++ src/components/Share/posters/index.tsx
@@ -1,0 +1,3 @@
+export { HolderPoster } from './holderPoster'
+export { IssuerPoster } from './issuerPoster'
+export { NftPoster } from './nftPoster'
