+++ src/views/Issuer/index.tsx
@@ -8,7 +8,7 @@ import { ReactComponent as ShareSvg } from '../../assets/svg/share.svg'
 import { Appbar, HEADER_HEIGHT } from '../../components/Appbar'
 import { useHistory } from 'react-router-dom'
 import { Share } from '../../components/Share'
-import { PosterType } from '../../components/Share/poster.interface'
+import { PosterType } from '../../components/Share/posters/poster.interface'
 import { useQuery } from 'react-query'
 import { Query } from '../../models'
 import { useWalletModel } from '../../hooks/useWallet'
